+++ jkit-sql/src/main/java/com/alianga/jkit/sql/SqlTokenType.java
@@ -277,6 +277,11 @@ public enum SqlTokenType {
     BEGIN,
     DECLARE,
     VACUUM,
+    COPY,
+    HANDLER,
+    PREPARE,
+    EXECUTE,
+    DEALLOCATE,
     GO;
 
     /**